Adjust clickable areas to highlight correctly
Adding a `cursor: pointer` CSS property to an element causes it to be
highlighted in some browsers when clicked multiple times. In this case
clicking the up/down arrows highlighted the entire parent frog
container, which was a bit confusing. This updates it to only highlight
the arrow that is being clicked on.